Update a project Sprint 3.2
