+++ README.md
@@ -3,48 +3,31 @@ locators.py
 # фикстуры
 conftest.py
 
-# проверка ввода имени при регистрации в системе Stellar Burgers
-test_01_1_registration_check_name.py
+# 1.проверка ввода данных при регистрации в системе Stellar Burgers включает в себя:
+# - проверку ввода имени при регистрации 
+# - проверку ввода email при регистрации 
+# - проверку ввода пароля при регистрации (менее 6 символов)
+# - проверку ввода пароля при регистрации (более 6 символов)
+test_01_registration_form.py
 
-# проверка ввода email при регистрации 
-test_01_2_registration_check_email.py
-
-# проверка ввода пароля при регистрации (более 6 символов)
-test_01_3_registration_check_password_more_6_sings.py
-
-# проверка ввода пароля при регистрации (менее 6 символов)
-test_01_4_registration_check_password_less_6_sings.py
-
+#  2. проверка авторизации:
 # проверка входа через кнопку в форме регистрации
-test_02_login_via_button_registration_form.py
-
 # проверка входа через кнопку «Личный кабинет» 
-test_03_enter_into_personal_account.py 
-
 # проверка входа по кнопке «Войти в аккаунт» на главной
-test_04_enter_into_log_in_to_the_account.py
-
 # проверка входа через кнопку в форме восстановления пароля
-test_05_login_via_the_button_recovery_form.py
+# проверка выхода из аккаунта
+test_02_authorization_in_the_system.py
 
-#  проверка перехода в личный кабинет
-test_06_move_to_the_personal_account.py
-
-#  проверка перехода из личного кабинета в "Конструктор"
-test_07_move_from_personal_account_to_constructor.py
-
+# 3. Перемещение по сайту
+# проверка перехода в личный кабинет
+# проверка перехода из личного кабинета в "Конструктор"
 # проверка перехода из личного кабинета на главную через "Логотип"
-test_08_move_from_personal_account_to_logo.py
 
-#  проверка выхода из аккаунта
-test_09_log_out_of_your_personal_account.py
+test_03_move_through_the_system.py
 
+# 4. Переключение по табам "Конструктора"
 #  проверка перехода из Раздела "Конструктор" к разделу "Начинки"
-test_10_move_from_constructor_to_fillings.py
-
 #  проверка перехода из Раздела "Конструктор" к разделу "Соусы"
-test_11_move_from_constructor_to_sauce.py
-
 #  проверка перехода из Раздела "Конструктор" к разделу "Булки"
-test_12_move_from_constructor_to_buns.py
+test_04_transfer_from_constructor_to_tabs.py
