--- tests/test_03_enter_into_personal_account.py
@@ -1,16 +1,0 @@
-from locators import TestLocators
-from selenium.webdriver.support.wait import WebDriverWait
-
-class TestLogin:
-    def test_login_via_personal_account(self, driver):
-        driver.get("https://stellarburgers.nomoreparties.site/")
-        personal_button = driver.find_elements(*TestLocators.Search_personal_button_main_locator)[0]
-        personal_button.click()
-        driver.get("https://stellarburgers.nomoreparties.site/login")
-        email = driver.find_element(*TestLocators.Input_email_locator).send_keys("user169@yandex.ru")
-        password = driver.find_element(*TestLocators.Input_password_locator).send_keys("Mir1941@")
-        driver.find_element(*TestLocators.Log_in_locator).click()
-        driver.get("https://stellarburgers.nomoreparties.site/")
-        assert driver.current_url == "https://stellarburgers.nomoreparties.site/"
-        WebDriverWait(driver, 10)
-
