--- tests/test_04_enter_into_log_in_to_the_account.py
@@ -1,20 +1,0 @@
-from locators import TestLocators
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.common.by import By
-
-class TestLogin:
-    def test_login_via_account(self, driver):
-        driver.get("https://stellarburgers.nomoreparties.site/")
-        driver.find_element(*TestLocators.Search_button_main_locator).click()
-        driver.get("https://stellarburgers.nomoreparties.site/login")
-        email = driver.find_elements(*TestLocators.Input_email_locator)[0]
-        email.send_keys("user169@yandex.ru")
-        WebDriverWait(driver, 10).until(expected_conditions.visibility_of_element_located((By.XPATH, ".//input[@value='user169@yandex.ru']")))
-        password = driver.find_element(*TestLocators.Input_password_locator).send_keys("Mir1941@")
-        WebDriverWait(driver, 10).until(expected_conditions.visibility_of_element_located((By.XPATH, ".//input[@value='Mir1941@']")))
-        driver.find_element(*TestLocators.Log_in_locator).click()
-        driver.get("https://stellarburgers.nomoreparties.site/")
-        assert driver.current_url == "https://stellarburgers.nomoreparties.site/"
-
-
